@@ -17,69 +17,81 @@
   "license": "MIT",
   "private": true,
   "devDependencies": {
-    "@babel/core": "^7.8.4",
-    "@fortawesome/fontawesome-svg-core": "^1.2.27",
-    "@fortawesome/free-brands-svg-icons": "^5.12.1",
-    "@fortawesome/free-solid-svg-icons": "^5.12.1",
-    "@fortawesome/react-fontawesome": "^0.1.8",
-    "@mdx-js/loader": "^1.5.5",
-    "@rollup/plugin-alias": "^3.0.1",
-    "@rollup/plugin-typescript": "^2.1.0",
-    "@types/common-tags": "^1.8.0",
-    "@types/jest": "^25.1.2",
-    "@types/lodash": "^4.14.149",
-    "@types/react": "^16.9.17",
-    "@types/react-copy-to-clipboard": "^4.3.0",
-    "@types/react-dom": "^16.9.5",
-    "@types/react-inspector": "^4.0.0",
-    "@types/react-router-dom": "^5.1.3",
-    "@types/react-test-renderer": "^16.9.2",
-    "@typescript-eslint/eslint-plugin": "^2.21.0",
-    "@typescript-eslint/parser": "^2.16.0",
-    "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.0.6",
-    "babel-preset-react-app": "^9.1.1",
-    "codesandbox": "^2.1.12",
-    "common-tags": "^1.8.0",
-    "delay": "^4.3.0",
-    "eslint": "^6.8.0",
-    "eslint-config-react-app": "^5.1.0",
-    "eslint-plugin-flowtype": "^3.13.0",
-    "eslint-plugin-import": "^2.20.0",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.17.0",
-    "eslint-plugin-react-hooks": "^1.7.0",
-    "glob": "^7.1.6",
-    "jest": "^25.1.0",
-    "lodash": "^4.17.15",
-    "md5": "^2.2.1",
-    "prettier": "^1.19.1",
-    "react": "^16.12.0",
-    "react-copy-to-clipboard": "^5.0.2",
-    "react-docgen-typescript-loader": "^3.6.0",
-    "react-dom": "^16.13.0",
-    "react-inspector": "^4.0.0",
-    "react-router-dom": "^5.1.2",
-    "react-test-renderer": "^16.13.0",
-    "rollup": "^1.32.0",
-    "ts-loader": "^6.2.1",
-    "typescript": "^3.7.5",
-    "webpack": "^4.41.6",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.10.3"
+    "@babel/core": "7.9.6",
+    "@babel/plugin-transform-runtime": "7.9.6",
+    "@babel/polyfill": "7.8.7",
+    "@fortawesome/fontawesome-svg-core": "1.2.28",
+    "@fortawesome/free-brands-svg-icons": "5.13.0",
+    "@fortawesome/free-solid-svg-icons": "5.13.0",
+    "@fortawesome/react-fontawesome": "0.1.9",
+    "@mdx-js/loader": "1.6.0",
+    "@rollup/plugin-babel": "5.0.0",
+    "@rollup/plugin-node-resolve": "7.1.3",
+    "@types/common-tags": "1.8.0",
+    "@types/jest": "25.2.1",
+    "@types/lodash": "4.14.150",
+    "@types/react": "16.9.34",
+    "@types/react-copy-to-clipboard": "4.3.0",
+    "@types/react-dom": "16.9.7",
+    "@types/react-inspector": "4.0.0",
+    "@types/react-router-dom": "5.1.5",
+    "@types/react-test-renderer": "16.9.2",
+    "@typescript-eslint/eslint-plugin": "2.30.0",
+    "@typescript-eslint/parser": "2.30.0",
+    "babel-eslint": "10.1.0",
+    "babel-loader": "8.1.0",
+    "babel-plugin-module-resolver": "4.0.0",
+    "babel-preset-react-app": "9.1.2",
+    "codesandbox": "2.1.14",
+    "common-tags": "1.8.0",
+    "css-loader": "3.5.3",
+    "delay": "4.3.0",
+    "eslint": "6.8.0",
+    "eslint-config-react-app": "5.2.1",
+    "eslint-plugin-flowtype": "4.7.0",
+    "eslint-plugin-import": "2.20.2",
+    "eslint-plugin-jsx-a11y": "6.2.3",
+    "eslint-plugin-react": "7.19.0",
+    "eslint-plugin-react-hooks": "3.0.0",
+    "glob": "7.1.6",
+    "jest": "25.5.3",
+    "lodash": "4.17.15",
+    "md5": "2.2.1",
+    "mini-css-extract-plugin": "0.9.0",
+    "normalize.css": "8.0.1",
+    "postcss-loader": "3.0.0",
+    "postcss-preset-env": "6.7.0",
+    "prettier": "2.0.5",
+    "react": "16.13.1",
+    "react-copy-to-clipboard": "5.0.2",
+    "react-docgen-typescript-loader": "3.7.2",
+    "react-dom": "16.13.1",
+    "react-inspector": "5.0.1",
+    "react-router-dom": "5.1.2",
+    "react-style-system": "0.1.0",
+    "react-test-renderer": "16.13.1",
+    "rollup": "2.7.6",
+    "style-loader": "1.2.1",
+    "sugarss": "2.0.0",
+    "ts-loader": "7.0.2",
+    "typescript": "3.8.3",
+    "webpack": "4.43.0",
+    "webpack-cli": "3.3.11",
+    "webpack-dev-server": "3.10.3"
   },
   "dependencies": {
-    "@types/classnames": "^2.2.9",
-    "@types/invariant": "^2.2.31",
-    "@types/shortid": "0.0.29",
-    "classnames": "^2.2.6",
-    "invariant": "^2.2.4",
-    "polished": "^3.4.4",
-    "shortid": "^2.2.15",
-    "stylis": "^3.5.4"
+    "@babel/runtime": "7.9.6",
+    "@types/classnames": "2.2.10",
+    "@types/invariant": "2.2.32",
+    "classnames": "2.2.6",
+    "nanoid": "3.1.4",
+    "polished": "3.6.1"
   },
   "peerDependencies": {
     "react": "^16.8.0",
-    "react-dom": "^16.8.0"
-  }
+    "react-dom": "^16.8.0",
+    "react-style-system": "0.1.0"
+  },
+  "sideEffects": false,
+  "browserslist": "> 5%"
 }
