@@ -1,9 +1,9 @@
-const typescript = require('@rollup/plugin-typescript');
+import babel from '@rollup/plugin-babel';
+import resolve from '@rollup/plugin-node-resolve';
 
 const input = 'src/index.ts';
-const plugins = [typescript()];
 
-module.exports = [
+export default [
   {
     input,
     output: {
@@ -11,7 +11,32 @@ module.exports = [
       format: 'esm',
       sourcemap: true,
     },
-    plugins,
+    plugins: [
+      resolve({
+        extensions: ['.ts', '.tsx', '.js', '.jsx'],
+        modulesOnly: true,
+      }),
+      babel({
+        babelrc: false,
+        babelHelpers: 'runtime',
+        extensions: ['.ts', '.tsx', '.js', '.jsx'],
+        presets: [
+          ['@babel/preset-env', { targets: '> 5% and not IE 11' }],
+          '@babel/preset-typescript',
+          '@babel/preset-react',
+        ],
+        plugins: ['@babel/plugin-transform-runtime'],
+      }),
+    ],
+    external: [
+      'react',
+      'classnames',
+      'polished',
+      'nanoid',
+      'react-dom',
+      'react-style-system',
+      /^@babel\/runtime/,
+    ],
   },
   {
     input,
@@ -20,7 +45,38 @@ module.exports = [
       format: 'umd',
       name: 'HackerUI',
       sourcemap: true,
+      globals: {
+        react: 'React',
+        classnames: 'classNames',
+        polished: 'polished',
+        'react-style-system': 'reactStyleSystem',
+        nanoid: 'nanoid',
+        'react-dom': 'ReactDOM',
+      },
     },
-    plugins,
+    plugins: [
+      resolve({
+        extensions: ['.ts', '.tsx', '.js', '.jsx'],
+        modulesOnly: true,
+      }),
+      babel({
+        babelrc: false,
+        babelHelpers: 'bundled',
+        extensions: ['.ts', '.tsx', '.js', '.jsx'],
+        presets: [
+          ['@babel/preset-env', { targets: 'defaults, not IE 11' }],
+          '@babel/preset-typescript',
+          '@babel/preset-react',
+        ],
+      }),
+    ],
+    external: [
+      'react',
+      'classnames',
+      'polished',
+      'nanoid',
+      'react-dom',
+      'react-style-system',
+    ],
   },
 ];
