@@ -1,14 +1,6 @@
 import React, { useState } from 'react';
-import {
-  createStyles,
-  PropsFromStyles,
-  Button,
-  TimesIcon,
-  useTheme,
-  Chip,
-  ChipThumbnail,
-  CheckIcon,
-} from 'hacker-ui';
+import { Button, TimesIcon, Chip, ChipThumbnail, CheckIcon } from 'hacker-ui';
+import { createStyles, PropsFromStyles, useTheme } from 'react-style-system';
 import { readableColor } from 'polished';
 
 const useStyles = createStyles(({ css, theme }) => ({
@@ -17,7 +9,7 @@ const useStyles = createStyles(({ css, theme }) => ({
   `,
   card: css`
     margin: ${theme.gap(1)} auto;
-    width: ${theme.breakpoints.mobile}px;
+    width: ${theme.breakpoints.mobile};
     max-width: 100%;
     display: flex;
     flex-direction: column;
@@ -97,7 +89,7 @@ function FilterChipsExample(props: Props) {
           <Button
             className={styles.closeButton}
             shape="icon"
-            on={theme.colors.brand}
+            surface={theme.colors.brand}
             color={readableColor(theme.colors.brand)}
           >
             <TimesIcon />
