@@ -1,8 +1,5 @@
 import React, { useState } from 'react';
 import {
-  useTheme,
-  createStyles,
-  PropsFromStyles,
   Button,
   Modal,
   ModalHeader,
@@ -10,6 +7,7 @@ import {
   ModalFooter,
   ModalActions,
 } from 'hacker-ui';
+import { createStyles, PropsFromStyles, useTheme } from 'react-style-system';
 
 const useStyles = createStyles(({ css, theme }) => ({
   root: css`
@@ -46,6 +44,8 @@ function ModalExample(props: Props) {
 
   const [open, setOpen] = useState(false);
 
+  console.log(styles.cssVariableObject);
+
   return (
     <>
       <Root>
@@ -58,13 +58,17 @@ function ModalExample(props: Props) {
         </Button>
       </Root>
 
-      <Modal open={open} onClose={() => setOpen(false)}>
+      <Modal
+        style={styles.cssVariableObject}
+        open={open}
+        onClose={() => setOpen(false)}
+      >
         <ModalHeader className={styles.modalHeader}>
           <h3 className={styles.title}>Modal Header</h3>
         </ModalHeader>
         <ModalContent>
           <div className={styles.modalScroll}>
-            {Array.from(Array(10)).map(i => (
+            {Array.from(Array(10)).map((_, i) => (
               <p key={i} className={styles.paragraph}>
                 Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
                 eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
