@@ -10,15 +10,16 @@ import {
   useLocation,
 } from 'react-router-dom';
 import {
-  useTheme,
-  createStyles,
-  PropsFromStyles,
   List,
   ListItem,
   ListItemButton,
+  useMediaQuery,
+  Drawer,
+  Button,
 } from 'hacker-ui';
+import { createStyles, PropsFromStyles, useTheme } from 'react-style-system';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faCaretDown } from '@fortawesome/free-solid-svg-icons';
+import { faCaretDown, faBars } from '@fortawesome/free-solid-svg-icons';
 
 // Styles
 const useStyles = createStyles(({ css, theme }) => {
@@ -34,6 +35,10 @@ const useStyles = createStyles(({ css, theme }) => {
       box-shadow: ${theme.shadows.standard};
       margin: ${theme.gap(1)};
       height: 500px;
+
+      ${theme.breakpoints.down(theme.breakpoints.tablet)} {
+        margin: ${theme.space(1)};
+      }
     `,
     title: css`
       ${theme.fonts.body1};
@@ -74,12 +79,20 @@ const useStyles = createStyles(({ css, theme }) => {
       display: flex;
       align-items: center;
     `,
+    openNavButton: css`
+      margin-right: ${theme.space(0.5)};
+    `,
     main: css`
       flex: 1 1 auto;
       display: flex;
       & > p {
         ${theme.fonts.h2};
         margin: auto;
+
+        ${theme.breakpoints.down(theme.breakpoints.tablet)} {
+          ${theme.fonts.h3};
+          margin: auto;
+        }
       }
     `,
     listItemButton: css`
@@ -159,10 +172,7 @@ const links: Links = [
 ];
 
 function titleToSlug(title: string) {
-  return title
-    .toLowerCase()
-    .trim()
-    .replace(/ /g, '-');
+  return title.toLowerCase().trim().replace(/ /g, '-');
 }
 
 /**
@@ -202,7 +212,7 @@ const toggle = (set: { [key: string]: boolean }, keyToToggle: string) =>
 
 const remove = (set: { [key: string]: boolean }, keyToRemove: string) =>
   Object.keys(set)
-    .filter(key => key !== keyToRemove)
+    .filter((key) => key !== keyToRemove)
     .reduce((set, key) => {
       set[key] = true;
       return set;
@@ -219,7 +229,11 @@ function NavExample(props: Props) {
   const { Root, styles } = useStyles(props);
   const theme = useTheme();
   const [collapsed, setCollapsed] = useState({} as { [key: string]: boolean });
+  const [drawerOpen, setDrawerOpen] = useState(false);
   const location = useLocation();
+  const isMobile = useMediaQuery(
+    theme.breakpoints.down(theme.breakpoints.tablet),
+  );
 
   /**
    * recursively creates
@@ -237,7 +251,7 @@ function NavExample(props: Props) {
 
         const handleClick = () => {
           if (!isFolder) return;
-          setCollapsed(collapsed => toggle(collapsed, path));
+          setCollapsed((collapsed) => toggle(collapsed, path));
         };
 
         return (
@@ -249,9 +263,8 @@ function NavExample(props: Props) {
                 })}
                 style={{ paddingLeft: theme.space(depth + 1) }}
                 onClick={handleClick}
-                component={
-                  isFolder ? 'button' : props => <Link to={path} {...props} />
-                }
+                component={isFolder ? 'button' : Link}
+                {...(isFolder ? undefined : { to: path })}
               >
                 <span className={styles.listItemText}>{title}</span>{' '}
                 {isFolder && (
@@ -272,24 +285,53 @@ function NavExample(props: Props) {
     </List>
   );
 
+  const navContent = (
+    <>
+      <div className={styles.title}>Nav Example</div>
+      {makeList(links)}
+    </>
+  );
+
   return (
-    <Root>
-      <nav className={styles.nav}>
-        <div className={styles.title}>Nav Example</div>
-        {makeList(links)}
-      </nav>
-      <div className={styles.content}>
-        <header className={styles.header}>{location.pathname}</header>
-        <main className={styles.main}>
-          <Switch>
-            {pages.map(({ path, component }) => (
-              <Route path={path} exact component={component} />
-            ))}
-            <Redirect to={firstPage.path} />
-          </Switch>
-        </main>
-      </div>
-    </Root>
+    <>
+      <Root>
+        {!isMobile && <nav className={styles.nav}>{navContent}</nav>}
+        <div className={styles.content}>
+          <header className={styles.header}>
+            {isMobile && (
+              <Button
+                className={styles.openNavButton}
+                shape="icon"
+                aria-label="Open Nav"
+                onClick={() => setDrawerOpen(true)}
+              >
+                <FontAwesomeIcon icon={faBars} size="lg" />
+              </Button>
+            )}
+            {location.pathname}
+          </header>
+          <main className={styles.main}>
+            <Switch>
+              {pages.map(({ path, component }) => (
+                <Route path={path} exact component={component} />
+              ))}
+              <Redirect to={firstPage.path} />
+            </Switch>
+          </main>
+        </div>
+      </Root>
+
+      {isMobile && (
+        <Drawer
+          style={styles.cssVariableObject}
+          className={styles.nav}
+          open={drawerOpen}
+          onClose={() => setDrawerOpen(false)}
+        >
+          {navContent}
+        </Drawer>
+      )}
+    </>
   );
 }
 
