@@ -1,11 +1,13 @@
 import React, { forwardRef } from 'react';
 import classNames from 'classnames';
-import { transparentize } from 'polished';
-import createStyles from './createStyles';
-import { ReactComponent, PropsFromStyles } from './types';
-import ColorProvider from './ColorProvider';
-import useColorContext from './useColorContext';
-import useTheme from './useTheme';
+import { transparentize, readableColor } from 'polished';
+import {
+  createStyles,
+  ColorContextProvider,
+  useColorContext,
+  PropsFromStyles,
+} from 'react-style-system';
+import { ReactComponent } from './types';
 
 const useStyles = createStyles(({ css, theme, color }) => ({
   // button base styles
@@ -36,25 +38,25 @@ const useStyles = createStyles(({ css, theme, color }) => ({
     }
 
     & > .hui-button {
-      margin-top: -${theme.space(0.5)};
-      margin-bottom: -${theme.space(0.5)};
-      margin-right: -${theme.space(0.75)};
-      margin-left: -${theme.space(0.25)};
+      margin-top: ${`-${theme.space(0.5)}`};
+      margin-bottom: ${`-${theme.space(0.5)}`};
+      margin-right: ${`-${theme.space(0.75)}`};
+      margin-left: ${`-${theme.space(0.25)}`};
     }
   `,
   // variants
   outlined: css`
-    border: 1px solid ${color.onSurface};
-    color: ${color.onSurface};
+    border: 1px solid ${color.readable};
+    color: ${color.readable};
 
-    transition: background-color ${theme.durations.standard}ms,
-      border ${theme.durations.standard}ms;
+    transition: background-color ${theme.durations.standard},
+      border ${theme.durations.standard};
 
     &:disabled {
       &,
       &:hover,
       &:focus {
-        color: ${transparentize(0.4, color.onSurface)};
+        color: ${transparentize(0.4, color.readable)};
         background-color: transparent;
       }
     }
@@ -62,39 +64,39 @@ const useStyles = createStyles(({ css, theme, color }) => ({
   outlinedClickable: css`
     cursor: pointer;
     &:focus {
-      background-color: ${transparentize(0.92, color.asBackground)};
+      background-color: ${transparentize(0.92, color.decorative)};
     }
     &:hover {
-      background-color: ${transparentize(0.9, color.asBackground)};
+      background-color: ${transparentize(0.9, color.decorative)};
     }
     &:active {
-      background-color: ${transparentize(0.8, color.asBackground)};
+      background-color: ${transparentize(0.8, color.decorative)};
     }
   `,
   filled: css`
-    background-color: ${color.asBackground};
-    color: ${color.bgContrast};
+    background-color: ${color.decorative};
+    color: ${readableColor(color.decorative)};
     border: 2px solid transparent;
-    transition: background-color ${theme.durations.standard}ms,
-      border ${theme.durations.standard}ms;
+    transition: background-color ${theme.durations.standard},
+      border ${theme.durations.standard};
 
     &:disabled {
-      background-color: ${transparentize(0.5, color.asBackground)};
+      background-color: ${transparentize(0.5, color.decorative)};
       border: 2px solid transparent;
-      color: ${transparentize(0.2, color.bgContrast)};
+      color: ${transparentize(0.2, readableColor(color.decorative))};
     }
   `,
   filledClickable: css`
     cursor: pointer;
     &:focus {
-      background-color: ${transparentize(0.18, color.asBackground)};
+      background-color: ${transparentize(0.18, color.decorative)};
       border: 2px solid ${transparentize(0.7, '#fff')};
     }
     &:hover {
-      background-color: ${transparentize(0.23, color.asBackground)};
+      background-color: ${transparentize(0.23, color.decorative)};
     }
     &:active {
-      background-color: ${transparentize(0.3, color.asBackground)};
+      background-color: ${transparentize(0.3, color.decorative)};
       border: 2px solid ${transparentize(0.5, '#fff')};
     }
   `,
@@ -122,18 +124,18 @@ const Chip = forwardRef((props: Props, ref: React.Ref<any>) => {
     ...restOfProps
   } = useStyles(props, component);
 
-  const theme = useTheme();
   const colorContext = useColorContext();
 
-  let color = props.color ?? colorContext?.on ?? theme.colors.accent;
-  let on = props.on ?? colorContext?.on ?? theme.colors.surface;
+  let color = props.color ?? colorContext?.color.original;
+  let surface = props.surface ?? colorContext?.surface;
 
   if (variant === 'filled') {
-    on = color;
+    surface = color;
+    color = readableColor(surface);
   }
 
   return (
-    <ColorProvider on={on} color={color}>
+    <ColorContextProvider surface={surface} color={color}>
       <Root
         className={classNames({
           [styles.outlined]: variant === 'outlined',
@@ -144,8 +146,10 @@ const Chip = forwardRef((props: Props, ref: React.Ref<any>) => {
         ref={ref}
         {...restOfProps}
       />
-    </ColorProvider>
+    </ColorContextProvider>
   );
 });
+
+Chip.displayName = 'Chip';
 
 export default Chip;
