@@ -1,6 +1,5 @@
 import React, { forwardRef } from 'react';
-import createStyles from './createStyles';
-import { PropsFromStyles } from './types';
+import { createStyles, PropsFromStyles } from 'react-style-system';
 
 const useStyles = createStyles(({ css, theme }) => ({
   root: css`
@@ -16,5 +15,7 @@ const ListItem = forwardRef((props: Props, ref: React.Ref<HTMLLIElement>) => {
   const { Root, styles, ...restOfProps } = useStyles(props, 'li');
   return <Root ref={ref} {...restOfProps} />;
 });
+
+ListItem.displayName = 'ListItem';
 
 export default ListItem;
