@@ -1,12 +1,12 @@
 import React, { forwardRef, useState, useEffect } from 'react';
 import { transparentize, readableColor } from 'polished';
 import { createPortal } from 'react-dom';
-import createStyles from './createStyles';
-import { PropsFromStyles, ReactComponent } from './types';
+import { createStyles, PropsFromStyles } from 'react-style-system';
+import { ReactComponent } from './types';
 
-const useStyles = createStyles(({ css, theme }) => ({
+const useStyles = createStyles(({ css, theme, staticVar }) => ({
   root: css`
-    max-height: 90vh;
+    max-height: 90%;
     width: ${theme.block(8)};
     max-width: 100%;
     background-color: ${theme.colors.surface};
@@ -16,13 +16,14 @@ const useStyles = createStyles(({ css, theme }) => ({
     overflow: hidden;
     display: flex;
     flex-direction: column;
+    z-index: ${theme.zIndex.modal};
   `,
   container: css`
     position: fixed;
-    width: 100vw;
-    height: 100vh;
     top: 0;
     left: 0;
+    bottom: 0;
+    right: 0;
     display: flex;
   `,
   backdrop: css`
@@ -31,7 +32,7 @@ const useStyles = createStyles(({ css, theme }) => ({
     left: 0;
     bottom: 0;
     right: 0;
-    background-color: ${transparentize(0.5, 'black')};
+    background-color: ${staticVar(transparentize(0.5, 'black'))};
     z-index: ${theme.zIndex.modal};
   `,
 }));
@@ -49,6 +50,7 @@ const Modal = forwardRef((props: Props, ref: React.Ref<HTMLElement>) => {
     Root,
     styles,
     open,
+    component: _component,
     // TODO: in the future, we should bind to the escape button for closing
     onClose,
     ...restOfProps
@@ -69,7 +71,14 @@ const Modal = forwardRef((props: Props, ref: React.Ref<HTMLElement>) => {
       document.body.removeChild(container);
       setContainer(null);
     };
-  }, [open, styles.container]);
+  }, [open, props.style, styles.container]);
+
+  useEffect(() => {
+    if (!container) return;
+    for (const [k, v] of Object.entries(styles.cssVariableObject)) {
+      container.style.setProperty(k, v);
+    }
+  }, [container, styles.cssVariableObject]);
 
   // TODO: throw something in the DOM later for SSR SEO
   return (
@@ -83,5 +92,7 @@ const Modal = forwardRef((props: Props, ref: React.Ref<HTMLElement>) => {
     )
   );
 });
+
+Modal.displayName = 'Modal';
 
 export default Modal;
