@@ -1,14 +1,16 @@
 import React, { forwardRef, useContext } from 'react';
 import classNames from 'classnames';
 import { transparentize } from 'polished';
-import createDynamicColorPalette from './createDynamicColorPalette';
+import {
+  createStyles,
+  createReadablePalette,
+  PropsFromStyles,
+} from 'react-style-system';
 import FormControlContext from './FormControlContext';
-import createStyles from './createStyles';
-import { PropsFromStyles } from './types';
 
-const useStyles = createStyles(({ css, color, theme, givenSurface }) => {
-  const bland = createDynamicColorPalette(theme.colors.bland, givenSurface);
-  const danger = createDynamicColorPalette(theme.colors.danger, givenSurface);
+const useStyles = createStyles(({ css, color, theme, surface }) => {
+  const bland = createReadablePalette(theme.colors.bland, surface);
+  const danger = createReadablePalette(theme.colors.danger, surface);
 
   const width = theme.space(3.5);
   const height = theme.space(2);
@@ -23,32 +25,32 @@ const useStyles = createStyles(({ css, color, theme, givenSurface }) => {
     root: css`
       position: relative;
       margin: ${theme.space(0.5)} 0;
-      background-color: ${givenSurface};
+      background-color: ${surface};
       border-radius: 99999px;
     `,
     hasError: css`
-      color: ${danger.asBackground};
+      color: ${danger.decorative};
       & .facade {
-        border: 2px solid ${danger.asBackground};
+        border: 2px solid ${danger.decorative};
       }
       & .switch:focus ~ .facade {
-        border: 2px solid ${danger.asBackground};
-        background-color: ${transparentize(0.93, danger.asBackground)};
+        border: 2px solid ${danger.decorative};
+        background-color: ${transparentize(0.93, danger.decorative)};
       }
       & .switch:not([disabled]):hover ~ .facade {
-        border: 2px solid ${transparentize(0.3, danger.asBackground)};
-        background-color: ${transparentize(0.93, danger.asBackground)};
+        border: 2px solid ${transparentize(0.3, danger.decorative)};
+        background-color: ${transparentize(0.93, danger.decorative)};
       }
       & .switch:active ~ .facade {
-        background-color: ${transparentize(0.9, danger.asBackground)};
+        background-color: ${transparentize(0.9, danger.decorative)};
       }
 
       & .switch:checked ~ .facade {
-        background-color: ${transparentize(0.93, danger.asBackground)};
+        background-color: ${transparentize(0.93, danger.decorative)};
       }
 
       & .switch:checked ~ .facade .dot {
-        fill: ${danger.asBackground};
+        fill: ${danger.decorative};
       }
     `,
     switch: css`
@@ -59,21 +61,21 @@ const useStyles = createStyles(({ css, color, theme, givenSurface }) => {
       left: 0;
 
       &:focus ~ .facade {
-        border: 2px solid ${color.asBackground};
-        background-color: ${transparentize(0.93, color.asBackground)};
+        border: 2px solid ${color.decorative};
+        background-color: ${transparentize(0.93, color.decorative)};
       }
       &:hover ~ .facade {
-        border: 2px solid ${transparentize(0.3, color.asBackground)};
-        background-color: ${transparentize(0.93, color.asBackground)};
+        border: 2px solid ${transparentize(0.3, color.decorative)};
+        background-color: ${transparentize(0.93, color.decorative)};
       }
       &:active ~ .facade {
-        background-color: ${transparentize(0.9, color.asBackground)};
+        background-color: ${transparentize(0.9, color.decorative)};
       }
 
       &:disabled ~ .facade {
         cursor: not-allowed;
-        border: 2px solid ${transparentize(0.7, bland.asBackground)};
-        background-color: ${transparentize(0.9, bland.asBackground)};
+        border: 2px solid ${transparentize(0.7, bland.decorative)};
+        background-color: ${transparentize(0.9, bland.decorative)};
       }
 
       &:disabled {
@@ -82,16 +84,16 @@ const useStyles = createStyles(({ css, color, theme, givenSurface }) => {
 
       &:checked ~ .facade .dot {
         cx: 125;
-        fill: ${color.asBackground};
+        fill: ${color.decorative};
       }
 
       &:disabled ~ .facade .dot {
-        fill: ${bland.asBackground};
+        fill: ${bland.decorative};
       }
 
       &:checked ~ .facade {
-        /* border: 2px solid ${bland.asBackground}; */
-        background-color: ${transparentize(0.9, color.asBackground)};
+        /* border: 2px solid ${bland.decorative}; */
+        background-color: ${transparentize(0.9, color.decorative)};
       }
     `,
     switchStandard: css`
@@ -115,10 +117,10 @@ const useStyles = createStyles(({ css, color, theme, givenSurface }) => {
       align-items: center;
       flex: 0 0 auto;
       background-color: white;
-      transition: border ${theme.durations.standard}ms,
-        background-color ${theme.durations.standard}ms;
-      border: 2px solid ${bland.asBackground};
-      background-color: ${givenSurface};
+      transition: border ${theme.durations.standard},
+        background-color ${theme.durations.standard};
+      border: 2px solid ${bland.decorative};
+      background-color: ${surface};
     `,
     facadeSmall: css`
       width: ${widthSmall};
@@ -139,8 +141,8 @@ const useStyles = createStyles(({ css, color, theme, givenSurface }) => {
       pointer-events: none;
     `,
     dot: css`
-      transition: cx ${theme.durations.standard}ms,
-        fill ${theme.durations.standard}ms;
+      transition: cx ${theme.durations.standard},
+        fill ${theme.durations.standard};
     `,
     dotSmall: css`
       r: 30;
@@ -245,5 +247,7 @@ const Switch = forwardRef((props: Props, ref: React.Ref<HTMLDivElement>) => {
     </Root>
   );
 });
+
+Switch.displayName = 'Switch';
 
 export default Switch;
