@@ -1,22 +1,57 @@
 const path = require('path');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 const include = [
   path.resolve(__dirname, './src'),
   path.resolve(__dirname, './docs'),
   path.resolve(__dirname, './website'),
   path.resolve(__dirname, './examples'),
+  path.resolve(__dirname, './.cache'),
 ];
 
 module.exports = {
-  mode: 'development',
+  mode: process.NODE_ENV === 'production' ? 'production' : 'development',
   entry: './website/index.tsx',
   output: {
     filename: 'hacker-ui-docs.js',
     publicPath: '/',
     path: path.resolve(__dirname, './public'),
+    chunkFilename: '[id].js',
   },
+  plugins: [
+    new MiniCssExtractPlugin({
+      filename: 'main.css',
+    }),
+  ],
   module: {
     rules: [
+      {
+        test: /\.rss-css$/,
+        use: [
+          MiniCssExtractPlugin.loader,
+          {
+            loader: 'css-loader',
+            options: { importLoaders: 2 },
+          },
+          'postcss-loader',
+          '@react-style-system/loader',
+        ],
+        include: [
+          ...include,
+          require.resolve('@react-style-system/loader/load.rss-css'),
+        ],
+        sideEffects: true,
+      },
+      {
+        test: /\.css$/,
+        use: [MiniCssExtractPlugin.loader, 'css-loader', 'postcss-loader'],
+        include: [
+          path.resolve(__dirname, './.cache'),
+          path.resolve(__dirname, './website'),
+        ],
+        exclude: [path.resolve(__dirname, './node_modules')],
+        sideEffects: true,
+      },
       {
         test: /\.doc-folder$/,
         include,
@@ -29,8 +64,15 @@ module.exports = {
         test: /\.example\.tsx$/,
         include,
         use: [
-          { loader: path.resolve(__dirname, './loaders/exampleLoader.js') },
+          {
+            loader: path.resolve(__dirname, './loaders/exampleLoader.js'),
+            options: {
+              cacheDir: path.resolve(__dirname, './.cache'),
+            },
+          },
         ],
+        // TODO: sideEffects: true needs to be audited
+        sideEffects: true,
       },
       {
         test: /\.mdx$/,
@@ -45,11 +87,13 @@ module.exports = {
           path.resolve(__dirname, './examples'),
         ],
         loader: 'babel-loader',
+        // TODO: sideEffects: true needs to be audited
+        sideEffects: true,
       },
     ],
   },
   resolve: {
-    extensions: ['.tsx', '.ts', '.js', '.jsx', '.mdx', '.doc-folder'],
+    extensions: ['.tsx', '.ts', '.js', '.jsx', '.mdx', '.doc-folder', '.css'],
     alias: {
       'hacker-ui': path.resolve(__dirname, './src/index.ts'),
       examples: path.resolve(__dirname, './examples'),
