@@ -1,13 +1,14 @@
-import React, { useEffect } from 'react';
+import React, { useEffect, useState, memo, Suspense } from 'react';
 import { useHistory, Switch, Route, Redirect } from 'react-router-dom';
-import { useCssReset, createStyles, StyleProps } from 'hacker-ui';
-import flattenDocArray from './flattenDocArray';
+import { createStyles, PropsFromStyles } from 'react-style-system';
 import docArray from '../docs';
+import flattenDocArray from './flattenDocArray';
 
 import Nav from './Nav';
 import AppBar from './AppBar';
 import NoRoute from './NoRoute';
 import PageWrapper from './PageWrapper';
+import LoadingView from './LoadingView';
 
 const routes = flattenDocArray(docArray).map(
   ({ component, ...restOfProps }) => {
@@ -52,44 +53,62 @@ const useStyles = createStyles(({ css, theme }) => ({
   `,
 }));
 
-interface Props extends StyleProps<typeof useStyles> {}
+interface Props extends PropsFromStyles<typeof useStyles> {}
 
 function App(props: Props) {
-  useCssReset();
   const { Root, styles } = useStyles(props);
+  const [mobileNavOpen, setMobileNavOpen] = useState(false);
 
   const history = useHistory();
 
   useEffect(() => {
-    const unsubscribe = history.listen(() => {
+    const handler = () =>
+      // TODO: this timeout is not reliable
       setTimeout(() => {
         const { Prism } = window as any;
         if (!Prism) return;
         Prism.highlightAll();
-      }, 0);
-    });
+
+        window.scrollTo(0, 0);
+      }, 500);
+
+    const unsubscribe = history.listen(handler);
+    handler();
 
     return unsubscribe;
   }, [history]);
 
   return (
     <Root>
-      <Nav className={styles.nav} />
+      <Nav
+        className={styles.nav}
+        open={mobileNavOpen}
+        onClose={() => setMobileNavOpen(false)}
+      />
       <div className={styles.content}>
-        <AppBar className={styles.appBar} />
+        <AppBar
+          className={styles.appBar}
+          onOpenMobileNav={() => setMobileNavOpen(true)}
+        />
         <main className={styles.main}>
-          <Switch>
-            <Route path="/" exact render={() => <Redirect to={firstPath} />} />
-            {routes.map(({ title, ...restOfRoute }, index) => (
-              <Route key={index} {...restOfRoute} />
-            ))}
-            <Route path="/404" component={NoRoute} />
-            <Redirect to="/404" />
-          </Switch>
+          <Suspense fallback={<LoadingView />}>
+            <Switch>
+              <Route
+                path="/"
+                exact
+                render={() => <Redirect to={firstPath} />}
+              />
+              {routes.map(({ title, ...restOfRoute }, index) => (
+                <Route key={index} {...restOfRoute} />
+              ))}
+              <Route path="/404" component={NoRoute} />
+              <Redirect to="/404" />
+            </Switch>
+          </Suspense>
         </main>
       </div>
     </Root>
   );
 }
 
-export default App;
+export default memo(App);
