@@ -2,19 +2,40 @@ import React, { useState, Fragment } from 'react';
 import classNames from 'classnames';
 import { darken, lighten, readableColor } from 'polished';
 import {
-  createStyles,
-  StyleProps,
   List,
   ListItemButton,
-  useTheme,
+  useMediaQuery,
+  Drawer,
+  Button,
+  TimesIcon,
+  ListItem,
+  Tooltip,
 } from 'hacker-ui';
-import { Link, useLocation } from 'react-router-dom';
-import docArray, { DocArray } from '../docs';
-import { toggle } from './setOperations';
+import { createStyles, PropsFromStyles, useTheme } from 'react-style-system';
+import { Link } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCaretDown } from '@fortawesome/free-solid-svg-icons';
+import docArray, { DocArray } from '../docs';
 // @ts-ignore
 import { version } from '../package.json';
+import { toggle } from './setOperations';
+
+function flattenDocArray(arr: DocArray): DocArray {
+  return arr
+    .map((i) => (Array.isArray(i.value) ? flattenDocArray(i.value) : [i]))
+    .reduce((flattened, next) => {
+      for (const i of next) {
+        flattened.push(i);
+      }
+      return flattened;
+    }, [] as DocArray);
+}
+
+const flattenedDocArray = docArray.map((i) =>
+  Array.isArray(i.value)
+    ? { title: i.title, value: flattenDocArray(i.value) }
+    : i,
+) as DocArray;
 
 const useStyles = createStyles(({ css, theme }) => {
   const backgroundColor =
@@ -30,12 +51,19 @@ const useStyles = createStyles(({ css, theme }) => {
       display: flex;
       flex-direction: column;
       overflow: hidden;
+      height: 100%;
     `,
     header: css`
       flex: 0 0 auto;
       padding: ${theme.space(1)};
       border-bottom: 1px solid ${theme.colors.bland};
       min-height: ${theme.block(0.75)};
+      display: flex;
+    `,
+    closeButton: css`
+      margin-right: ${theme.space(1)};
+    `,
+    headerInfo: css`
       display: flex;
       flex-direction: column;
       justify-content: center;
@@ -68,101 +96,122 @@ const useStyles = createStyles(({ css, theme }) => {
     routeBody: css`
       ${theme.fonts.caption};
     `,
-    itemActive: css`
-      background-color: ${darken(0.09, backgroundColor)};
-
-      &:focus {
-        background-color: ${darken(0.07, backgroundColor)};
-      }
-      &:hover {
-        background-color: ${darken(0.05, backgroundColor)};
-      }
-      &:active {
-        background-color: ${darken(0.03, backgroundColor)};
-      }
-    `,
   };
 });
 
-interface Props extends StyleProps<typeof useStyles> {}
+interface Props extends PropsFromStyles<typeof useStyles> {
+  open: boolean;
+  onClose: () => void;
+}
 
 function Nav(props: Props) {
-  const { Root, styles } = useStyles(props, 'aside');
-  const location = useLocation();
+  const { Root, styles, open, onClose } = useStyles(props, 'aside');
   const theme = useTheme();
   const [collapsedSet, setCollapsedSet] = useState<{ [key: string]: boolean }>(
     {},
   );
 
-  const makeList = (
-    docArray: DocArray,
-    rootPath = '',
-    depth = 0,
-  ): React.ReactNode => {
-    return (
-      <Fragment key={`root-${rootPath}`}>
-        {docArray.map(({ title, value }) => {
-          const slug = title.toLowerCase().replace(/ /g, '-');
-          const path = `${rootPath}/${slug}`;
-          const isDirectory = Array.isArray(value);
-          const isCollapsed = collapsedSet[path];
-          const active = location.pathname === path;
+  const isMobile = useMediaQuery(
+    theme.breakpoints.down(theme.breakpoints.tablet),
+  );
 
-          const handleClick = (path: string) => {
-            if (!isDirectory) return;
-            setCollapsedSet(set => toggle(set, path));
-          };
-
-          return (
-            <li key={`sub-${path}`}>
-              <ListItemButton
-                className={classNames(styles.item, {
-                  [styles.itemActive]: active,
-                })}
-                style={{ paddingLeft: theme.space(depth + 1) }}
-                component={
-                  isDirectory
-                    ? 'button'
-                    : (props: any) => <Link to={path} {...props} />
-                }
-                onClick={() => handleClick(path)}
-              >
-                <div
-                  className={classNames(styles.itemTitle, {
-                    [styles.itemTitleBold]: isDirectory,
-                  })}
-                >
-                  {title}
-                </div>
-                {isDirectory && (
-                  <FontAwesomeIcon
-                    className={styles.icon}
-                    icon={faCaretDown}
-                    rotation={isCollapsed ? 90 : undefined}
-                  />
-                )}
-              </ListItemButton>
-
-              {!isCollapsed && Array.isArray(value) && (
-                <List>{makeList(value, path, depth + 1)}</List>
-              )}
-            </li>
-          );
-        })}
-      </Fragment>
-    );
-  };
-
-  return (
+  const content = (
     <Root>
       <div className={styles.header}>
-        <h1 className={styles.title}>Hacker UI</h1>
-        <span className={styles.version}>{version}</span>
+        {isMobile && (
+          <Tooltip title="Close Nav" position="right">
+            {(tooltipProps) => (
+              <Button
+                {...tooltipProps}
+                className={styles.closeButton}
+                shape="icon"
+                color={theme.colors.brand}
+                aria-label="Close Nav"
+                onClick={onClose}
+                autoFocus
+              >
+                <TimesIcon />
+              </Button>
+            )}
+          </Tooltip>
+        )}
+        <div className={styles.headerInfo}>
+          <h1 className={styles.title}>
+            <span>Hacker UI</span>
+          </h1>
+          <span className={styles.version}>{version}</span>
+        </div>
       </div>
       <nav className={styles.body}>
-        <List>{makeList(docArray)}</List>
+        <List>
+          {flattenedDocArray.map(({ title, value }) => {
+            const rootSlug = `/${title.toLowerCase().replace(/ /g, '-')}`;
+
+            if (!Array.isArray(value)) {
+              throw new Error('Document must be in a folder');
+            }
+
+            return (
+              <Fragment key={rootSlug}>
+                <ListItem>
+                  <ListItemButton
+                    className={styles.item}
+                    onClick={() =>
+                      setCollapsedSet((set) => toggle(set, rootSlug))
+                    }
+                  >
+                    <div
+                      className={classNames(
+                        styles.itemTitle,
+                        styles.itemTitleBold,
+                      )}
+                    >
+                      {title}
+                    </div>
+                    <FontAwesomeIcon
+                      className={styles.icon}
+                      icon={faCaretDown}
+                      rotation={collapsedSet[rootSlug] ? 90 : undefined}
+                    />
+                  </ListItemButton>
+                </ListItem>
+
+                {!collapsedSet[rootSlug] && (
+                  <List>
+                    {value.map(({ title }) => {
+                      const slug = title.toLowerCase().replace(/ /g, '-');
+                      const path = `${rootSlug}/${slug}`;
+
+                      return (
+                        <ListItem key={slug}>
+                          <ListItemButton
+                            className={styles.item}
+                            component={Link}
+                            // @ts-ignore
+                            to={path}
+                            onClick={onClose}
+                          >
+                            <div className={styles.itemTitle}>{title}</div>
+                          </ListItemButton>
+                        </ListItem>
+                      );
+                    })}
+                  </List>
+                )}
+              </Fragment>
+            );
+          })}
+        </List>
       </nav>
     </Root>
+  );
+
+  if (!isMobile) return content;
+
+  return (
+    <Drawer open={open} onClose={onClose}>
+      {content}
+    </Drawer>
   );
 }
 
