@@ -1,5 +1,5 @@
 import React from 'react';
-import { createStyles, StyleProps } from 'hacker-ui';
+import { createStyles, StyleProps } from 'react-style-system';
 
 const useStyles = createStyles(({ css }) => ({
   root: css``,
@@ -8,7 +8,7 @@ const useStyles = createStyles(({ css }) => ({
 interface Props extends StyleProps<typeof useStyles> {}
 
 function NoRoute(props: Props) {
-  const { Root, styles } = useStyles(props);
+  const { Root } = useStyles(props);
 
   return <Root>404 not found</Root>;
 }
