@@ -1,10 +1,14 @@
 import React from 'react';
-import { createStyles, PropsFromStyles } from 'hacker-ui';
+import { createStyles, PropsFromStyles } from 'react-style-system';
 import { transparentize } from 'polished';
 
 const useStyles = createStyles(({ css, theme }) => ({
   root: css`
     padding: ${theme.gap(1)};
+    ${theme.breakpoints.down(theme.breakpoints.tablet)} {
+      padding: ${theme.space(1)};
+    }
+
     width: 1024px;
     max-width: 100%;
     margin: 0 auto;
@@ -12,14 +16,27 @@ const useStyles = createStyles(({ css, theme }) => ({
 
     & > h1 {
       ${theme.fonts.h3};
+
+      ${theme.breakpoints.down(theme.breakpoints.tablet)} {
+        ${theme.fonts.h4};
+      }
     }
 
     & > h2 {
       ${theme.fonts.h4};
+
+      ${theme.breakpoints.down(theme.breakpoints.tablet)} {
+        ${theme.fonts.h5};
+      }
     }
 
     & > h3 {
       ${theme.fonts.h5};
+
+      ${theme.breakpoints.down(theme.breakpoints.tablet)} {
+        ${theme.fonts.body1};
+        font-weight: 500;
+      }
     }
 
     & > p {
@@ -31,6 +48,7 @@ const useStyles = createStyles(({ css, theme }) => ({
       background-color: #2d2d2d;
       padding: ${theme.space(1)};
       color: white;
+      font-size: 0.8rem;
 
       & code {
         background-color: transparent;
